requirements.txt
python-telegram-bot>=21.0
httpx>=0.27
python-decouple>=3.8
cachetools>=5.3
